Move model and result types into Messages lib
