+++ src/AIDemo.Web/Entities/Contact.cs
@@ -1,5 +1,5 @@
 using System.ComponentModel.DataAnnotations.Schema;
-using System.Globalization;
+using AIDemo.Web.Messages;
 using Google.Apis.PeopleService.v1.Data;
 using Microsoft.EntityFrameworkCore;
 using NetTopologySuite.Geometries;
@@ -110,21 +110,11 @@ public record ContactAddress(
             location);
 };
 
-[Owned]
-public record ContactBirthday(int Year, int Month, int Day)
+
+public static class ContactExtensions
 {
-    public static ContactBirthday? FromGoogleContactBirthday(Birthday birthday) =>
-        birthday.Date is { } birthDate
-            ? new(birthDate.Year ?? 0, birthDate.Month ?? 0, birthDate.Day ?? 0)
-            : null;
-
-    public override string ToString() =>
-        (Year, Month, Day) switch
-        {
-            (_, 0, 0) => "Unknown",
-            (0, _, 0) => CultureInfo.CurrentCulture.DateTimeFormat.GetMonthName(Month), // Get month name from Month number
-            (0, _, _) => $"{CultureInfo.CurrentCulture.DateTimeFormat.GetMonthName(Month)} {Day}",
-            _ => $"{CultureInfo.CurrentCulture.DateTimeFormat.GetMonthName(Month)} {Day} {Year}",
-        };
-};
-
+    public static ContactBirthdayResult ToContactBirthdayResult(this Contact contact) =>
+        new(
+            contact.Names.First().UnstructuredName,
+            contact.Birthday ?? new ContactBirthday(0, 0, 0));
+}
