+++ src/AIDemo.Web/Eventing/Consumers/ContactImportConsumer.cs
@@ -1,3 +1,4 @@
+using AIDemo.Web.Messages;
 using Google.Apis.PeopleService.v1.Data;
 using Haack.AIDemoWeb.Entities;
 using Haack.AIDemoWeb.Library;
